+++ test/visualTest.ts
@@ -24,22 +24,21 @@
  *  THE SOFTWARE.
  */
 
-// import * as d3 from "d3";
 import powerbiVisualsApi from "powerbi-visuals-api";
 import DataView = powerbiVisualsApi.DataView;
-import VisualObjectInstance = powerbiVisualsApi.VisualObjectInstance;
 import { valueFormatter as vf } from "powerbi-visuals-utils-formattingutils";
-import { assertColorsMatch, d3Click} from "powerbi-visuals-utils-testutils";
-// import valueFormatter = vf.valueFormatter;
+import { assertColorsMatch, d3Click } from "powerbi-visuals-utils-testutils";
 import { VisualData } from "./visualData";
 import { VisualBuilder } from "./visualBuilder";
 import { SunburstDataPoint } from "../src/dataInterfaces";
 
 const DefaultWaitForRender: number = 500;
 const LegendSelector: string = "#legendGroup";
+const SliceSelector: string = ".sunburst__slice";
 const SliceLabelSelector: string = ".sunburst__slice-label";
 const LabelVisibleSelector: string = ".sunburst__label--visible";
 const PercentageSelector: string = ".sunburst__percentage-label";
+const CategoryLabelSelector: string = ".sunburst__category-label";
 
 describe("Sunburst", () => {
     let visualBuilder: VisualBuilder;
@@ -110,101 +109,74 @@ describe("Sunburst", () => {
 
             visualBuilder.updateRenderTimeout(
                 dataView,
-                () => {
-                    const firstPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length-1];
+                async () => {
+                    const firstPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
                     const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
                     d3Click(firstPoint, 5, 5);
-                    setTimeout(
-                        () => {
-                            expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(2);
-                            d3Click(secondClickPoint, 1, 1);
-                            setTimeout(
-                                () => {
-                                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
-                                    done();
-                                },
-                                DefaultWaitForRender);
-                        },
-                        DefaultWaitForRender);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(2);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
+                    done();
                 },
                 2,
                 DefaultWaitForRender);
         });
 
-        it("category labels should be hidden", (done: DoneFn) => {
+        beforeEach(() => {
             dataView = defaultDataViewBuilder.getDataView(
                 [
                     defaultDataViewBuilder.RegionsDataSet,
                     defaultDataViewBuilder.CountriesDataSet
                 ]);
+        });
 
+        it("category labels should be hidden", (done: DoneFn) => {
             dataView.metadata.objects = {
                 group: { showSelected: false }
             };
 
             visualBuilder.updateRenderTimeout(
                 dataView,
-                () => {
+                async () => {
                     const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
                     const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
                     d3Click(firstClickPoint, 5, 5);
-                    setTimeout(
-                        () => {
-                            expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(1);
-                            d3Click(secondClickPoint, 1, 1);
-                            setTimeout(
-                                () => {
-                                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
-                                    done();
-                                },
-                                DefaultWaitForRender);
-                        },
-                        DefaultWaitForRender);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(1);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
+                    done();
                 },
                 2,
                 DefaultWaitForRender);
         });
 
         it("category labels should be visible always", (done: DoneFn) => {
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
-
             dataView.metadata.objects = {
                 group: { showSelected: true }
             };
 
             visualBuilder.updateRenderTimeout(
                 dataView,
-                () => {
+                async () => {
                     const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
                     const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
                     d3Click(firstClickPoint, 5, 5);
-                    setTimeout(
-                        () => {
-                            expect($(LabelVisibleSelector).length).toBe(2);
-                            d3Click(secondClickPoint, 1, 1);
-                            setTimeout(
-                                () => {
-                                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
-                                    done();
-                                },
-                                DefaultWaitForRender);
-                        },
-                        DefaultWaitForRender);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(2);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(0);
+                    done();
                 },
                 2,
                 DefaultWaitForRender);
         });
 
         it("data labels should not be hidden by default", (done: DoneFn) => {
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
             visualBuilder.updateRenderTimeout(
                 dataView,
                 () => {
@@ -214,11 +186,6 @@ describe("Sunburst", () => {
         });
 
         it("count of data labels should be equal slice count", (done: DoneFn) => {
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
             dataView.metadata.objects = {
                 group: { showDataLabels: true }
             };
@@ -229,6 +196,116 @@ describe("Sunburst", () => {
                     done();
                 }, 2, DefaultWaitForRender);
         });
+
+        it("data label font size should be correct", (done: DoneFn) => {
+            const fontSize: number = 22;
+            const expectedFontSize: string = "22px";
+            dataView.metadata.objects = {
+                group: { labelFontSize: fontSize }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(SliceLabelSelector)).style.fontSize).toBe(expectedFontSize);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("data label font weight should be correct", (done: DoneFn) => {
+            const fontWeight: boolean = true;
+            const expectedWeight: string = "bold";
+            dataView.metadata.objects = {
+                group: { labelFontBold: fontWeight }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(SliceLabelSelector)).style.fontWeight).toBe(expectedWeight);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("data label text decoration should be correct", (done: DoneFn) => {
+            const textDecoration: boolean = true;
+            const expectedDecoration: string = "underline";
+            dataView.metadata.objects = {
+                group: { labelFontUnderline: textDecoration }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(SliceLabelSelector)).style.textDecoration).toBe(expectedDecoration);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("data label font style should be correct", (done: DoneFn) => {
+            const fontItalic: boolean = true;
+            const expectedStyle: string = "italic";
+            dataView.metadata.objects = {
+                group: { labelFontItalic: fontItalic }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(SliceLabelSelector)).style.fontStyle).toBe(expectedStyle);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("label font family should be correct", (done: DoneFn) => {
+            const fontFamily: string = "Arial";
+            const expectedFamily: string = "Arial";
+            dataView.metadata.objects = {
+                group: { labelFontFamily: fontFamily }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(SliceLabelSelector)).style.fontFamily).toBe(expectedFamily);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
     });
 
     describe("Test invalid input data", () => {
@@ -237,7 +314,7 @@ describe("Sunburst", () => {
                 [
                     defaultDataViewBuilder.RegionsDataSet
                 ]);
-            dataView.matrix.rows.root.children = [];
+            dataView.matrix!.rows.root.children = [];
 
             visualBuilder.updateRenderTimeout(
                 dataView,
@@ -251,12 +328,15 @@ describe("Sunburst", () => {
     });
 
     describe("Legend", () => {
-        it("legend should be hidden by default", (done: DoneFn) => {
+        beforeEach(() => {
             dataView = defaultDataViewBuilder.getDataView(
                 [
                     defaultDataViewBuilder.RegionsDataSet,
                     defaultDataViewBuilder.CountriesDataSet
                 ]);
+        });
+
+        it("legend should be hidden by default", (done: DoneFn) => {
             visualBuilder.updateRenderTimeout(
                 dataView,
                 () => {
@@ -265,65 +345,96 @@ describe("Sunburst", () => {
                 }, 2, DefaultWaitForRender);
         });
         it("legend should be shown", (done: DoneFn) => {
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
             dataView.metadata.objects = {
                 legend: { show: true }
             };
             visualBuilder.updateRenderTimeout(
                 dataView,
                 () => {
-                    debugger;
                     expect(visualBuilder.element.querySelectorAll(`${LegendSelector.trim()}>*`).length).toBeTruthy();
                     done();
                 }, 2, DefaultWaitForRender);
         });
+        it("legend font family should be correct", (done: DoneFn) => {
+            const fontFamily: string = "Arial";
+            const expectedFamily: string = "Arial";
+            dataView.metadata.objects = {
+                legend: { show: true, fontFamily: fontFamily }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    for (const element of visualBuilder.element.querySelectorAll(`${LegendSelector.trim()} text`)) {
+                        expect((<HTMLElement>element).style.fontFamily).toBe(expectedFamily);
+                    }
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+        it("legend font size should be correct", (done: DoneFn) => {
+            const fontSize: number = 22;
+            const expectedSize: string = Math.round(fontSize * 4 / 3 * 10000) / 10000 + "px";
+            dataView.metadata.objects = {
+                legend: { show: true, fontSize: fontSize }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    for (const element of visualBuilder.element.querySelectorAll(`${LegendSelector.trim()} text`)) {
+                        expect((<HTMLElement>element).style.fontSize).toBe(expectedSize);
+                    }
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
     });
 
     describe("Central caption", () => {
-        it("percentage font size should be correct", (done: DoneFn) => {
+        beforeEach(() => {
             dataView = defaultDataViewBuilder.getDataView(
                 [
                     defaultDataViewBuilder.RegionsDataSet,
                     defaultDataViewBuilder.CountriesDataSet
                 ]);
+        });
 
+        it("percentage font size should be correct", (done: DoneFn) => {
             dataView.metadata.objects = {
                 group: { showSelected: false }
             };
 
             visualBuilder.updateRenderTimeout(
                 dataView,
-                () => {
+                async () => {
                     const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
                     const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
                     d3Click(firstClickPoint, 5, 5);
-                    setTimeout(
-                        () => {
-                            expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(1);
-                            d3Click(secondClickPoint, 1, 1);
-                            setTimeout(
-                                () => {
-                                    expect((<HTMLElement>visualBuilder.element.querySelector(PercentageSelector)).style['font-size']).toBe("28px");
-                                    done();
-                                },
-                                DefaultWaitForRender);
-                        },
-                        DefaultWaitForRender);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.element.querySelectorAll(LabelVisibleSelector).length).toBe(1);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(PercentageSelector)).style['font-size']).toBe("28px");
+                    done();
                 },
                 2,
                 DefaultWaitForRender);
         });
-        
+
         it("label font size should be correct", (done: DoneFn) => {
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
             const fontSize: number = 22;
             const expectedFontSize: string = "44px";
             dataView.metadata.objects = {
@@ -331,21 +442,103 @@ describe("Sunburst", () => {
             };
             visualBuilder.updateRenderTimeout(
                 dataView,
-                () => {
+                async () => {
                     const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
                     const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
                     d3Click(firstClickPoint, 5, 5);
-                    setTimeout(
-                        () => {
-                            d3Click(secondClickPoint, 1, 1);
-                            setTimeout(
-                                () => {
-                                    expect((<HTMLElement>visualBuilder.element.querySelector( PercentageSelector )).style.fontSize).toBe(expectedFontSize);
-                                    done();
-                                },
-                                DefaultWaitForRender);
-                        },
-                        DefaultWaitForRender);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(PercentageSelector)).style.fontSize).toBe(expectedFontSize);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("label font weight should be correct", (done: DoneFn) => {
+            const fontWeight: boolean = true;
+            const expectedWeight: string = "bold";
+            dataView.metadata.objects = {
+                group: { fontBold: fontWeight }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(CategoryLabelSelector)).style.fontWeight).toBe(expectedWeight);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("label text decoration should be correct", (done: DoneFn) => {
+            const textDecoration: boolean = true;
+            const expectedDecoration: string = "underline";
+            dataView.metadata.objects = {
+                group: { fontUnderline: textDecoration }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(CategoryLabelSelector)).style.textDecoration).toBe(expectedDecoration);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("label font style should be correct", (done: DoneFn) => {
+            const fontItalic: boolean = true;
+            const expectedStyle: string = "italic";
+            dataView.metadata.objects = {
+                group: { fontItalic: fontItalic }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(CategoryLabelSelector)).style.fontStyle).toBe(expectedStyle);
+                    done();
+                },
+                2,
+                DefaultWaitForRender);
+        });
+
+        it("label font family should be correct", (done: DoneFn) => {
+            const fontFamily: string = "Arial";
+            const expectedFamily: string = "Arial";
+            dataView.metadata.objects = {
+                group: { fontFamily: fontFamily }
+            };
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                async () => {
+                    const firstClickPoint: HTMLElement = visualBuilder.slices[visualBuilder.slices.length - 1];
+                    const secondClickPoint: HTMLElement = visualBuilder.mainElement[0];
+                    d3Click(firstClickPoint, 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    d3Click(secondClickPoint, 1, 1);
+                    await timeout(DefaultWaitForRender);
+                    expect((<HTMLElement>visualBuilder.element.querySelector(CategoryLabelSelector)).style.fontFamily).toBe(expectedFamily);
+                    done();
                 },
                 2,
                 DefaultWaitForRender);
@@ -353,14 +546,16 @@ describe("Sunburst", () => {
     });
 
     describe("Colors", () => {
-        it("should be parsed correctly", (done: DoneFn) => {
-            const color: string = "#006400";
+        beforeEach(() => {
             dataView = defaultDataViewBuilder.getDataView(
                 [
                     defaultDataViewBuilder.RegionsDataSet,
                     defaultDataViewBuilder.CountriesDataSet
                 ]);
-            dataView.matrix.rows.root.children[0].objects = {
+        });
+        it("should be parsed correctly", (done: DoneFn) => {
+            const color: string = "#006400";
+            dataView.matrix!.rows.root.children![0].objects = {
                 group: {
                     fill: {
                         solid: {
@@ -372,13 +567,13 @@ describe("Sunburst", () => {
             visualBuilder.updateRenderTimeout(
                 dataView,
                 () => {
-                    const result: VisualObjectInstance[] = visualBuilder.enumerateObjectInstances({ objectName: "group" });
-                    const colorExist: boolean = result.some((instance: VisualObjectInstance) =>
-                        instance.properties &&
-                        instance.properties["fill"] &&
-                        instance.properties["fill"]["solid"] &&
-                        instance.properties["fill"]["solid"]["color"] &&
-                        instance.properties["fill"]["solid"]["color"] === color);
+                    const result: powerbiVisualsApi.visuals.FormattingModel = visualBuilder.instance.getFormattingModel();
+                    const visual_card: powerbiVisualsApi.visuals.FormattingCard = <powerbiVisualsApi.visuals.FormattingCard>result.cards[0];
+                    const group_group: powerbiVisualsApi.visuals.FormattingGroup = <powerbiVisualsApi.visuals.FormattingGroup>visual_card.groups[0];
+                    const group_slices: powerbiVisualsApi.visuals.FormattingSlice[] = <powerbiVisualsApi.visuals.FormattingSlice[]>group_group.slices;
+                    const colorExist: boolean = group_slices.some((slice: powerbiVisualsApi.visuals.FormattingSlice) =>
+                        (<any>slice).control?.properties?.value?.value === color
+                    );
                     expect(colorExist).toBeTruthy();
                     done();
                 }, 2, DefaultWaitForRender);
@@ -388,13 +583,7 @@ describe("Sunburst", () => {
             const color: string = "#006400";
             const colorAsRGB: string = "rgb(0, 100, 0)";
 
-            dataView = defaultDataViewBuilder.getDataView(
-                [
-                    defaultDataViewBuilder.RegionsDataSet,
-                    defaultDataViewBuilder.CountriesDataSet
-                ]);
-
-            dataView.matrix.rows.root.children[0].objects = {
+            dataView.matrix!.rows.root.children![0].objects = {
                 group: {
                     fill: {
                         solid: {
@@ -488,7 +677,7 @@ describe("Sunburst", () => {
                 visualBuilder.updateRenderTimeout(dataView, () => {
                     const layers = visualBuilder.slices;
 
-                    expect(isColorAppliedToElements(layers, null, "fill"));
+                    expect(isColorAppliedToElements(layers, undefined, "fill"));
 
                     done();
                 });
@@ -520,6 +709,174 @@ describe("Sunburst", () => {
                 });
             }
         });
+
+        describe("keyboard navigation and related aria-attributes", () => {
+            let originalTimeout;
+
+            beforeEach(() => {
+                originalTimeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
+                jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
+                dataView = defaultDataViewBuilder.getDataView([
+                    defaultDataViewBuilder.RegionsDataSet,
+                    defaultDataViewBuilder.CountriesDataSet,
+                ]);
+            });
+
+            it("should have role=listbox and aria-multiselectable attributes correctly set", (done) => {
+                visualBuilder.updateRenderTimeout(dataView, () => {
+                    const gElement: HTMLElement = visualBuilder.element.firstChild?.firstChild?.firstChild as HTMLElement;
+
+                    expect(gElement.getAttribute("role")).toBe("listbox");
+                    expect(gElement.getAttribute("aria-multiselectable")).toBe("true");
+
+                    done();
+                }, DefaultWaitForRender);
+            });
+
+            it("should have tabindex, role=option, aria-label, and aria=selected correctly initialized", (done) => {
+                visualBuilder.updateRenderTimeout(dataView, () => {
+                    const slices = visualBuilder.slices;
+                    const [, ...categoryLabels] = JSON.stringify(dataView.matrix?.rows.root.children).split('"levelValues":[{"value":"').map((x: string) => x.replace("\\\"", "~").split('"')[0].replace("~", "\\\""));
+                    for (let i = 0; i < slices.length; i++) {
+                        const slice = slices[i];
+                        expect(slice.getAttribute("tabindex")).toBe("0");
+                        expect(slice.getAttribute("role")).toBe("option");
+                        expect(categoryLabels).toContain(slice.getAttribute("aria-label")!);
+                        expect(slice.getAttribute("aria-selected")).toBe("false");
+                    }
+
+                    done();
+                }, DefaultWaitForRender);
+            });
+
+            it("should have role=presentation correctly set on text labels", (done) => {
+                visualBuilder.updateRenderTimeout(dataView, () => {
+                    const gElement: HTMLElement = visualBuilder.element.firstChild?.firstChild?.firstChild as HTMLElement;
+                    const slices: NodeListOf<SVGPathElement> = gElement.querySelectorAll(SliceLabelSelector);
+                    for (const slice of slices) { 
+                        expect(slice.getAttribute("role")).toBe("presentation");
+                    }
+
+                    done();
+                }, DefaultWaitForRender);
+            });
+
+            it("aria attributs work when clicked", (done: DoneFn) => {
+                visualBuilder.updateRenderTimeout(dataView, async () => {
+                    d3Click(visualBuilder.slices[0], 5, 5);
+                    await timeout(DefaultWaitForRender);
+                    expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                    expect(visualBuilder.slices[0].style.opacity).toBe("1");
+                    for (const slice of visualBuilder.slices) {
+                        if (slice !== visualBuilder.slices[0]) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                            expect(slice.style.opacity).toBe("0.2");
+                        }
+                    }
+
+                    done();
+                });
+            });
+
+            it("enter toggles the correct slice", (done: DoneFn) => {
+                const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
+                visualBuilder.updateRenderTimeout(
+                    dataView,
+                    async () => {
+                        visualBuilder.slices[0].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                        expect(visualBuilder.slices[0].style.opacity).toBe("1");
+                        for (const slice of visualBuilder.slices) {
+                            if (slice !== visualBuilder.slices[0]) {
+                                expect(slice.getAttribute("aria-selected")).toBe("false");
+                                expect(slice.style.opacity).toBe("0.2");
+                            }
+                        }
+
+                        visualBuilder.slices[0].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        for (const slice of visualBuilder.slices) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                            expect(slice.style.opacity).toBe("1");
+                        }
+
+                        done();
+                    },
+                    2,
+                    DefaultWaitForRender);
+            });
+
+            it("space toggles the correct slice", (done: DoneFn) => {
+                const spaceEvent = new KeyboardEvent("keydown", { code: "Space", bubbles: true });
+                visualBuilder.updateRenderTimeout(
+                    dataView,
+                    async () => {
+                        visualBuilder.slices[0].dispatchEvent(spaceEvent);
+                        await timeout(DefaultWaitForRender);
+                        expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                        expect(visualBuilder.slices[0].style.opacity).toBe("1");
+                        for (const slice of visualBuilder.slices) {
+                            if (slice !== visualBuilder.slices[0]) {
+                                expect(slice.getAttribute("aria-selected")).toBe("false");
+                                expect(slice.style.opacity).toBe("0.2");
+                            }
+                        }
+
+                        visualBuilder.slices[0].dispatchEvent(spaceEvent);
+                        await timeout(DefaultWaitForRender);
+                        for (const slice of visualBuilder.slices) {
+                            expect(slice.getAttribute("aria-selected")).toBe("false");
+                            expect(slice.style.opacity).toBe("1");
+                        }
+
+                        done();
+                    },
+                    2,
+                    DefaultWaitForRender);
+            });
+
+            it("tab between slices works", (done: DoneFn) => {
+                const tabEvent = new KeyboardEvent("keydown", { code: "Tab", bubbles: true });
+                const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
+                visualBuilder.updateRenderTimeout(
+                    dataView,
+                    async () => {
+                        visualBuilder.slices[0].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        expect(visualBuilder.slices[0].getAttribute("aria-selected")).toBe("true");
+                        expect(visualBuilder.slices[0].style.opacity).toBe("1");
+                        for (const slice of visualBuilder.slices) {
+                            if (slice !== visualBuilder.slices[0]) {
+                                expect(slice.getAttribute("aria-selected")).toBe("false");
+                                expect(slice.style.opacity).toBe("0.2");
+                            }
+                        }
+
+                        visualBuilder.element.dispatchEvent(tabEvent);
+                        await timeout(DefaultWaitForRender);
+
+                        visualBuilder.slices[1].dispatchEvent(enterEvent);
+                        await timeout(DefaultWaitForRender);
+                        expect(visualBuilder.slices[1].getAttribute("aria-selected")).toBe("true");
+                        expect(visualBuilder.slices[1].style.opacity).toBe("1");
+                        for (const slice of visualBuilder.slices) {
+                            if (slice !== visualBuilder.slices[1]) {
+                                expect(slice.getAttribute("aria-selected")).toBe("false");
+                                expect(slice.style.opacity).toBe("0.2");
+                            }
+                        }
+
+                        done();
+                    },
+                    2,
+                    DefaultWaitForRender);
+            });
+
+            afterEach(() => {
+                jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
+            });
+        });
     });
 
     describe("covertTreeNodeToSunBurstDataPoint", () => {
@@ -531,8 +888,9 @@ describe("Sunburst", () => {
                 [],
                 {
                     dataPoints: [],
-                    total: 0,
-                    root: undefined,
+                    total: 0, 
+                    // @ts-ignore
+                    root: undefined, 
                 },
                 "#00ff00",
                 visualBuilder.visualHost,
@@ -545,3 +903,7 @@ describe("Sunburst", () => {
         });
     });
 });
+
+function timeout(ms: number) {
+    return new Promise(resolve => setTimeout(resolve, ms));
+}