+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 2.5.0
+* keyboard navigation (feature)
+* focus outline (styles)
+* aria roles (aria-selected, listbox, option) on slices
+
+## 2.4.0
+* Updated the deprecatedEnumerateInstances API to instead use the FormattingModel API (to go with Visual API version 5.1+)
+* refactored formatting model to use formatting settings service
+* Clarified some UI to make it more intuitive and expanded font size inputs to full font control inputs (font size, family, style, etc.).
+
+## 2.3.0
+* Fixed linting
+* Updated config in .stylelintrc following the CHANGELOG and migration guide
+* Added es-lint following the power-bi standards
+* Removed tslint
+* Fixed the code and styles to pass the linting
+
 ## 2.2.2
 * Updated visualData to use matrix builder from test utils
 * Updated test.webpack.config.json for development mode